booking details decomposed into separate files
